add multiday events > grid items are in place - hover is not there yet
